+++ js/src/address/index.ts
@@ -1,0 +1,4 @@
+export * from "../primary-domain";
+export * from "../utils/getSnsDomainKeysForOwner";
+export * from "../utils/getSnsDomainsForOwner";
+export * from "../utils/getSnsNftsForOwner";