+++ js/src/record/getRecord.ts
@@ -1,11 +1,13 @@
-import { Record as SnsRecord, Validation } from "@bonfida/sns-records";
+import { Buffer } from "buffer";
+
+import { Record as SnsRecord } from "@bonfida/sns-records";
 import { Connection, PublicKey } from "@solana/web3.js";
 
 import { NameRegistryState } from "../state";
 import { Record } from "../types/record";
 import { assertTldSupported } from "../utils/assertTldSupported";
 import { getSnsDomainKeySync } from "../utils/getSnsDomainKeySync";
-import { ETH_ROA_RECORDS, GUARDIANS, SELF_SIGNED } from "./const";
+import { ETH_ROA_RECORDS, GUARDIANS, SELF_SIGNED, Validation } from "./const";
 import { deserializeRecordContent } from "./deserializeRecordContent";
 import { getRecordV2Key } from "./getRecordV2Key";
 
@@ -15,12 +17,24 @@ interface GetRecordOptions {
 
 export interface RecordResult {
   record: Record;
-  retrievedRecord: SnsRecord;
+  retrievedRecord: RetrievedRecord;
   verified: {
     staleness: boolean;
     roa?: boolean;
   };
   deserializedContent?: string;
+}
+
+export interface RetrievedRecord {
+  header: {
+    stalenessValidation: number;
+    rightOfAssociationValidation: number;
+    contentLength: number;
+  };
+  data: Buffer;
+  getContent(): Buffer;
+  getStalenessId(): Buffer;
+  getRoAId(): Buffer;
 }
 
 /**