+++ js/src/record/verifyRightOfAssociation.ts
@@ -1,10 +1,10 @@
-import { Record as SnsRecord, Validation } from "@bonfida/sns-records";
+import { Record as SnsRecord } from "@bonfida/sns-records";
 import { Connection } from "@solana/web3.js";
 
 import { MissingVerifierError } from "../error";
 import { Record } from "../types/record";
 import { assertTldSupported } from "../utils/assertTldSupported";
-import { ETH_ROA_RECORDS, GUARDIANS } from "./const";
+import { ETH_ROA_RECORDS, GUARDIANS, Validation } from "./const";
 import { getRecordV2Key } from "./getRecordV2Key";
 
 import type { Buffer } from "buffer";