+++ js/src/types/index.ts
@@ -1,0 +1,2 @@
+export * from "./custom-bg";
+export * from "./record";